Send undefined instead of null (#297)
In #296 we made it so that we set empty object for empty responses.
Turns out the problem was with the body that we're sending. We coalesce
to null for undefined objects, but we should really just send undefined.
The backend does a check for Content-Length == 0, which is not true for
Null (Content-Length == 4), but is true for undefined.