Add pluggable activity context (#288)
This PR introduces the ability for plugins to add _typed_ values to the
activity context. It does this by adding a dedicated
`buildActivityContext` function that gets called in `app.process`. Then
via a bunch of generics plumbing we merge the plugin's context values
with the default ones.

We can use this to separate graph out into other packages, but keep the
contextual objects for the activity-handlers unchanged (aka. `appGraph`,
and `userGraph` will still be available). However, `graph` object on
`app` would most likely be part of the the `GraphPlugin` - (maybe we
make that accessible via a new `app.getPlugin('graph')` method?)

Here's a mini demo:


https://github.com/user-attachments/assets/328517e4-99b2-459d-ad05-80291c7b7e99





#### PR Dependency Tree


* **PR #288** 👈

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)