Simplify graph function paramDefs type (#335)
This PR changes the `paramDefs` type in the endpoint functions. The
`paramDefs` values are used to construct the Graph HTTP request - for
each parameter that the caller may supply, `paramDefs` decides whether
they go in the path, the query string, or the request header.

The `paramDefs` type so far has been: `{ name: string; in: 'query' |
'header' | 'path' }>[]`. So - an array of objects, each has a parameter
name, and an `in` value to show where it goes.

The new type is: `Partial<Record<'query' | 'header' | 'path',
string[]>>;` I'm pulling the 'in' values up a level to be a key in an
object, the values being the parameters for that 'in'.

In practice, the paramDefs for an API before may have been:
```
    paramDefs: [
      { name: '$top', in: 'query' },
      { name: '$skip', in: 'query' },
      { name: '$search', in: 'query' },
      { name: '$filter', in: 'query' },
      { name: '$count', in: 'query' },
      { name: '$orderby', in: 'query' },
      { name: '$select', in: 'query' },
      { name: '$expand', in: 'query' },
    ],
```
And now the same is:
```
    paramDefs: {
      query: ['$top', '$skip', '$search', '$filter', '$count', '$orderby', '$select', '$expand'],
    },
```

This change has a surprisingly large impact on the size of code
generated. After building the beta package, looking at the combined size
of the .js files, they shrink from 6207kB to 5204kB. That's about 1/6th
less code on disk.

There's no much improvement to the NPM package size, just a few kB,
since the removed repetition compresses so well. But the uncompressed
footprint makes it worthwhile all the same.


This change won't build until all endpoints are rebuilt, so in the first
iteration I included only the changes to the generation script & the
graph.call() method and a few updated endpoints - just enough to
showcase the change.

In the second iteration, I regenerated all the endpoints for both beta
and v.10 (schema commits 4e63218 and 3bdcc26, both published yesterday).